DAO implemented

+++ src/main/resources/db/changelog/changes/202608040002_create_order_items.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS order_items (
     id            UUID PRIMARY KEY,
     order_id      UUID NOT NULL,
     item_id       UUID NOT NULL,
-    quantity      INTEGER NOT NULL,
+    quantity      BIGINT NOT NULL,
     item_price    NUMERIC(19, 2) NOT NULL,
     created_at    TIMESTAMP    NOT NULL,
     updated_at    TIMESTAMP    NOT NULL,
